security check

+++ back-end/src/main/java/vn/hust/soict/lib/repository/ReceiptRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import vn.hust.soict.lib.entity.Receipt;
 @Repository
-public interface ReceiptRepository extends JpaRepository<Receipt, Integer> {
+public interface ReceiptRepository extends JpaRepository<Receipt, Long> {
 }
